Migrate Mutation.updateNativeFederation
