+++ packages/services/api/src/modules/schema/resolvers.ts
@@ -131,21 +131,6 @@ function __isTypeOf<
 
 export const resolvers: SchemaModule.Resolvers = {
   Mutation: {
-    async updateNativeFederation(_, { input }, { injector }) {
-      const translator = injector.get(IdTranslator);
-      const [organization, project] = await Promise.all([
-        translator.translateOrganizationId(input),
-        translator.translateProjectId(input),
-      ]);
-
-      return {
-        ok: await injector.get(SchemaManager).updateNativeSchemaComposition({
-          project,
-          organization,
-          enabled: input.enabled,
-        }),
-      };
-    },
     async updateProjectRegistryModel(_, { input }, { injector }) {
       const translator = injector.get(IdTranslator);
       const [organization, project] = await Promise.all([
